WIP: event_log table + updated query
This commit includes the spec for an event_log table in Tables.json. This needs the latest version of burrow which supports log table creation. Marking as WIP for now until the new burrow version is ready, and can be added to Dockerfile. Also includes an update to getAgreementData query to return ownerDisplayName.

Signed-off-by: Sam Sinha <ssinha484@gmail.com>